Rename FileEnumerator to FileScanner
