+++ src/core/FileScanner.h
